+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Release Notes for Siteseer
 
+## 2026-03-25
+- added `./craft siteseer/default/visit` command.
+- Added ddev config resources and added readme instructions for cached domain testing.
+- Made scheme/protocol configurable
+
 ## 1.0.0
 - Initial release